pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

# ────────────────────────────────────────────────────────────────────────────
# Core project metadata
# ────────────────────────────────────────────────────────────────────────────
[project]
name="DataTalks"
version = "0.0.2"               # ← bump when you tag a release
description = "Multimodal routing chatbot to Metro Areena (nbdev project)"
authors     = [{name = "Your Name", email = "you@example.com"}]
readme      = "README.md"
requires-python=">=3.9"

# Runtime deps (only what the package needs to *run*, not to *develop*)
dependencies = [
  "httpx>=0.26",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "polyline>=2.0"
]

[project.optional-dependencies]
dev = [
  "nbdev>=2.3",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff", "black",
  "httpx", "pydantic",
  "python-dotenv",
  "fastmcp",
  "agents",
  "python-fasthtml"
]

# ────────────────────────────────────────────────────────────────────────────
# nbdev / setuptools tweaks (kept from the original scaffold)
# ────────────────────────────────────────────────────────────────────────────
[tool.nbdev]
nbs_path = "nbs"
lib_path = "DataTalks"
doc_path = "_proc"

[tool.setuptools.package-data]

DataTalks = ["*.md", "*.html", "*.js", "*.css", "*.json"]

[tool.setuptools]
package-dir = {"" = "."}

